requirements.txt
pymongo~=4.6.1
requests~=2.31.0
python-dotenv~=1.0.1
Jinja2~=3.1.3
uvicorn~=0.27.1
APScheduler~=3.10.4
fastapi~=0.109.2
pytz~=2024.1